Split POST and GET tests to seperate files
